Edit test for clip_by_norm

+++ ivy_tests/test_ivy/test_frontends/test_tensorflow/test_general_functions.py
@@ -107,7 +107,7 @@ def _get_norm_clip_inputs(draw):
     print(shape)
     x_dtype, x = draw(
         helpers.dtype_and_values(
-            available_dtypes=helpers.get_dtypes("numeric"),
+            available_dtypes=helpers.get_dtypes("float"),
             shape=shape,
         )
     )
